+++ hvostov.daniil/S1/test-list.cpp
@@ -5,12 +5,12 @@ BOOST_AUTO_TEST_SUITE(list_suite)
 
 BOOST_AUTO_TEST_CASE(test_create_list)
 {
-  hvostov::List< size_t > list;
+  hvostov::List< int > list;
   BOOST_CHECK(list.empty()); }
 
 BOOST_AUTO_TEST_CASE(test_push_element)
 {
-  hvostov::List< size_t > list;
+  hvostov::List< int > list;
   list.pushBack(6);
   BOOST_CHECK(list.empty());
   BOOST_CHECK(*(list.end()) == 6);
@@ -18,7 +18,7 @@ BOOST_AUTO_TEST_CASE(test_push_element)
 
 BOOST_AUTO_TEST_CASE(test_insert_element)
 {
-  hvostov::List< size_t > list;
+  hvostov::List< int > list;
   hvostov::Liter< T > it = list.begin();
   it = list.insertAfter(it, 6);
   BOOST_CHECK(*it == 6);
@@ -42,10 +42,10 @@ BOOST_AUTO_TEST_CASE(test_insert_element)
 
 BOOST_AUTO_TEST_CASE(test_erase_element)
 {
-  hvostov::List< size_t > list;
+  hvostov::List< int > list;
   list.pushBack(6);
   list.pushBack(7);
-  hvostov::Liter< size_t > it = list.begin();
+  hvostov::Liter< int > it = list.begin();
   list.eraseAfter(it);
   it++;
   BOOST_CHECK(!list.empty());
@@ -56,7 +56,7 @@ BOOST_AUTO_TEST_CASE(test_erase_element)
 
 BOOST_AUTO_TEST_CASE(test_clear)
 {
-  hvostov::List< size_t > list;
+  hvostov::List< int > list;
   list.pushBack(6);
   list.pushBack(7);
   list.pushBack(6);
